Fix size narrowing size_t to int in FFmpegAudio/VideoDecoder
Use base::checked_cast<int> when assigning buffer_span.size() to
packet->size in FFmpegAudioDecoder::FFmpegDecode and
FFmpegVideoDecoder::FFmpegDecode to prevent safe narrowing overflow.

TAG=agy
CONV=4eee7864-0106-476f-b647-78eba7f8c730

Fixed: 536443672
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/8127445
Commit-Queue: Thomas Guilbert <tguilbert@chromium.org>
Commit-Queue: Dale Curtis <dalecurtis@chromium.org>
Auto-Submit: Dale Curtis <dalecurtis@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1665945}

+++ media/filters/ffmpeg_audio_decoder.cc
@@ -14,6 +14,7 @@ FASTVC-BENCH-CORPUS:chromium-main-v1-943b94ae-1c74-4335-94fa-ceb4d277cea8
 #include "base/feature_list.h"
 #include "base/functional/bind.h"
 #include "base/functional/callback_helpers.h"
+#include "base/numerics/safe_conversions.h"
 #include "base/task/sequenced_task_runner.h"
 #include "base/task/single_thread_task_runner.h"
 #include "media/base/audio_buffer.h"
@@ -226,7 +227,7 @@ bool FFmpegAudioDecoder::FFmpegDecode(const DecoderBuffer& buffer) {
   } else {
     auto buffer_span = base::span(buffer);
     packet->data = const_cast<uint8_t*>(buffer_span.data());
-    packet->size = buffer_span.size();
+    packet->size = base::checked_cast<int>(buffer_span.size());
     packet->pts =
         ConvertToTimeBase(codec_context_->time_base, buffer.timestamp());
